# doc/openapi/public.json
{
  "components": {
    "schemas": {
...
      "ErrorEntry": {
        "description": "One entry of `errors`/`warnings` in the envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error or warning code; use this rather than matching message text.",
            "type": "string"
          },
          "message": {
            "description": "Human-readable detail; messages may be in Portuguese.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "Failure": {
        "description": "Error response; data and warnings are absent.",
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "errors"
        ],
        "type": "object"
      },
...
      "Success_Vec_String": {
        "description": "Successful JSON response. Requests do not use this envelope.",
        "properties": {
          "data": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "warnings": {
            "description": "Nonfatal issues, omitted when absent (currently unknown_team on run ingestion).",
            "items": {
              "$ref": "#/components/schemas/ErrorEntry"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "wuerges@gmail.com",
      "name": "Emilio Wuerges"
    },
...
    "license": {
      "name": ""
    },
    "title": "Animeitor public API",
    "version": "2.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/events": {
      "get": {
        "description": "Event identifiers in creation order, including pre-start events. No authentication.",
        "operationId": "list_public_events",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    "regional-2026"
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/Success_Vec_String"
                }
              }
            },
            "description": "Successful result in data"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "Stored data is corrupt or uses an unsupported schema (storage_error)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "Database unavailable or busy (storage_unavailable). A failed write may have committed: read the resource before retrying."
          }
        },
        "summary": "List event identifiers",
        "tags": [
          "Public discovery"
        ]
      }
    },
    "/api/events/{event_name}/contests": {
      "get": {
        "description": "Alphabetically sorted contest names, available before and after event start so the landing page can link to upcoming countdowns. Unlike the internal list this returns strings, not configurations. An event with no contests returns an empty list. Contest state, configuration, and runs remain unavailable before start.",
        "operationId": "list_public_contests",
        "parameters": [
          {
            "description": "Event identifier (not a display label). URL-encode as a path segment.",
            "in": "path",
            "name": "event_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    "brasil"
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/Success_Vec_String"
                }
              }
            },
            "description": "Successful result in data"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "errors": [
                    {
                      "code": "not_found",
                      "message": "Required event, contest, or site does not exist."
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "Required event, contest, or site does not exist."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "Stored data is corrupt or uses an unsupported schema (storage_error)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "Database unavailable or busy (storage_unavailable). A failed write may have committed: read the resource before retrying."
          }
        },
        "summary": "List contest identifiers",
        "tags": [
          "Public discovery"
        ]
      }
    },
...
  }
}